abstracted bookingCard and refactored time string
